Modified for four ESCs
