+++ examples/DShotLibraryTest/DShotLibraryTest.ino
@@ -1,4 +1,4 @@
-#include <DShot.h>
+#include "DShot.h"
 
 /*
 
@@ -20,31 +20,31 @@ void setup() {
   Serial.begin(115200);
 
   // Notice, all pins must be connected to same PORT
-  esc1.attach(7);  
+  esc1.attach(7);
   esc1.setThrottle(throttle);
 }
 
 void loop() {
-  if (Serial.available()>0){
+  if (Serial.available() > 0) {
     target = Serial.parseInt();
-    if (target>2047)
-      target = 2047;
+    if (target > 2047) target = 2047;
     Serial.print(target, HEX);
     Serial.print("\t");
   }
-  if (throttle<48){
+  if (throttle < 48) {
     throttle = 48;
   }
-  if (target<=48){
+  if (target <= 48) {
     esc1.setThrottle(target);
-  }else{
-    if (target>throttle){
-      throttle ++;
+  } else {
+    if (target > throttle) {
+      throttle++;
       esc1.setThrottle(throttle);
-    }else if (target<throttle){
-      throttle --;
+    } else if (target < throttle) {
+      throttle--;
       esc1.setThrottle(throttle);
     }
   }
+
   delay(10);
 }
